[bknote71] bug fix

+++ src/main/java/com/bknote71/codecraft/robocode/leaderboard/LeaderBoardTemplate.java
@@ -23,12 +23,12 @@ public class LeaderBoardTemplate {
         this.ops = redisTemplate.opsForZSet();
     }
 
-    public static void updateLeaderBoard(int battleId, String username, int score) {
+    public static void updateLeaderBoard(int battleId, String username, double score) {
         System.out.println("update leader board: " + battleId + ", username" + username + ", score: " + score);
         ops.add(prefix + ":" + battleId, username, score); // 실시간 점수
     }
 
-    public static void updateTodayLeaderBoard(String username, int score) { // 죽었을 때나 혹은 종료되었을 때 업데이트
+    public static void updateTodayLeaderBoard(String username, double score) { // 죽었을 때나 혹은 종료되었을 때 업데이트
         ops.add(prefix + ":" + LocalDate.now().toString(), username, score);
     }
